Add PublicDatabase
Use Database.Factory to instantiate public and private database
instance.
