+++ skygear/src/androidTest/java/io/skygear/skygear/RecordDeleteRequestUnitTest.java
@@ -25,7 +25,7 @@ public class RecordDeleteRequestUnitTest {
     public static void setUpClass() throws Exception {
         instrumentationContext = InstrumentationRegistry.getContext().getApplicationContext();
         instrumentationContainer = new Container(instrumentationContext, Configuration.defaultConfiguration());
-        instrumentationPublicDatabase= Database.publicDatabase(instrumentationContainer);
+        instrumentationPublicDatabase= Database.Factory.publicDatabase(instrumentationContainer);
     }
 
     @AfterClass
