+++ skygear/src/main/java/io/skygear/skygear/Container.java
@@ -16,7 +16,7 @@ public final class Container {
     final Context context;
     final Pubsub pubsub;
     final RequestManager requestManager;
-    final Database publicDatabase;
+    final PublicDatabase publicDatabase;
     final Database privateDatabase;
     Configuration config;
 
@@ -34,8 +34,8 @@ public final class Container {
         this.requestManager = new RequestManager(context, config);
         this.pubsub = new Pubsub(this);
         this.persistentStore = new PersistentStore(context);
-        this.publicDatabase = Database.publicDatabase(this);
-        this.privateDatabase = Database.privateDatabase(this);
+        this.publicDatabase = Database.Factory.publicDatabase(this);
+        this.privateDatabase = Database.Factory.privateDatabase(this);
 
         this.auth = new AuthContainer(this);
 
@@ -242,73 +242,6 @@ public final class Container {
 
             this.requestManager.sendRequest(request);
         }
-    }
-
-    /**
-     * Gets default access control.
-     *
-     * @return the access control
-     */
-    public AccessControl getDefaultAccessControl() {
-        return AccessControl.defaultAccessControl();
-    }
-
-    /**
-     * Sets default access control.
-     *
-     * @param accessControl the access control
-     */
-    public void setDefaultAccessControl(AccessControl accessControl) {
-        this.persistentStore.defaultAccessControl = accessControl;
-        this.persistentStore.save();
-
-        AccessControl.defaultAccessControl = accessControl;
-    }
-
-    /**
-     * Sets admin roles.
-     *
-     * @param role    the roles array
-     * @param handler the handler
-     */
-    public void setAdminRole(Role[] role, SetRoleResponseHandler handler) {
-        SetAdminRoleRequest request = new SetAdminRoleRequest(role);
-        request.responseHandler = handler;
-
-        this.requestManager.sendRequest(request);
-    }
-
-    /**
-     * Sets an admin role.
-     *
-     * @param role    the role
-     * @param handler the handler
-     */
-    public void setAdminRole(Role role, SetRoleResponseHandler handler) {
-        this.setAdminRole(new Role[] { role }, handler);
-    }
-
-    /**
-     * Sets default roles.
-     *
-     * @param role    the role array
-     * @param handler the handler
-     */
-    public void setDefaultRole(Role[] role, SetRoleResponseHandler handler) {
-        SetDefaultRoleRequest request = new SetDefaultRoleRequest(role);
-        request.responseHandler = handler;
-
-        this.requestManager.sendRequest(request);
-    }
-
-    /**
-     * Sets a default role.
-     *
-     * @param role    the role
-     * @param handler the handler
-     */
-    public void setDefaultRole(Role role, SetRoleResponseHandler handler) {
-        this.setDefaultRole(new Role[] { role }, handler);
     }
 
     /**
