+++ skygear/src/main/java/io/skygear/skygear/PublicDatabase.java
@@ -1,0 +1,94 @@
+package io.skygear.skygear;
+
+
+/**
+ * The Skygear Public Database.
+ * <p>
+ * This class wraps the logic of public database concept in Skygear.
+ * In adition to record CRUD, this class also wraps record access api.
+ * </p>
+ */
+public class PublicDatabase extends Database {
+    /**
+     * Instantiates a new Database.
+     * <p>
+     * Please be reminded that the skygear container passed in would be weakly referenced.
+     * </p>
+     *
+     * @param databaseName the database name
+     * @param container    the container
+     */
+    PublicDatabase(String databaseName, Container container) {
+        super(databaseName, container);
+    }
+
+    /**
+     * Gets default access control.
+     *
+     * @return the access control
+     */
+    public AccessControl getDefaultAccessControl() {
+        return AccessControl.defaultAccessControl();
+    }
+
+    /**
+     * Sets default access control.
+     *
+     * @param accessControl the access control
+     */
+    public void setDefaultAccessControl(AccessControl accessControl) {
+        PersistentStore persistentStore = this.getContainer().persistentStore;
+        persistentStore.defaultAccessControl = accessControl;
+        persistentStore.save();
+
+        AccessControl.defaultAccessControl = accessControl;
+    }
+
+    /**
+     * Sets admin roles.
+     *
+     * @param role    the roles array
+     * @param handler the handler
+     */
+    public void setAdminRole(Role[] role, SetRoleResponseHandler handler) {
+        SetAdminRoleRequest request = new SetAdminRoleRequest(role);
+        request.responseHandler = handler;
+
+        RequestManager requestManager = this.getContainer().requestManager;
+        requestManager.sendRequest(request);
+    }
+
+    /**
+     * Sets an admin role.
+     *
+     * @param role    the role
+     * @param handler the handler
+     */
+    public void setAdminRole(Role role, SetRoleResponseHandler handler) {
+        this.setAdminRole(new Role[] { role }, handler);
+    }
+
+    /**
+     * Sets default roles.
+     *
+     * @param role    the role array
+     * @param handler the handler
+     */
+    public void setDefaultRole(Role[] role, SetRoleResponseHandler handler) {
+        SetDefaultRoleRequest request = new SetDefaultRoleRequest(role);
+        request.responseHandler = handler;
+
+        RequestManager requestManager = this.getContainer().requestManager;
+        requestManager.sendRequest(request);
+    }
+
+    /**
+     * Sets a default role.
+     *
+     * @param role    the role
+     * @param handler the handler
+     */
+    public void setDefaultRole(Role role, SetRoleResponseHandler handler) {
+        this.setDefaultRole(new Role[] { role }, handler);
+    }
+}
